agent/requirements.txt
python-dotenv==1.1.1
redis==6.4.0
pyyaml==6.0.2
livekit-agents==1.2.5
livekit-plugins-openai==1.2.5
livekit-plugins-cartesia==1.2.5
livekit-plugins-elevenlabs==1.2.5
livekit-plugins-deepgram==1.2.5
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-silero==1.2.5
livekit-plugins-hume==1.2.5

# Scheduling and task management
rq==2.4.1
rq-scheduler==0.14.0

# Medical knowledge storage uses Redis (no additional dependencies needed)

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0
freezegun==1.5.5

# Time zone handling
pytz==2025.2

# CLI tools
tabulate==0.9.0